Type a lowered temp by its value, not by its register
A float4 of texture offsets was declared int4: fxc reused r0.x for the
loop counter after the loop, and the finalizer typed a variable by the
register it was in. It types it by its value now, the way the hoist
already did - the readers where they agree, the operation that made it
otherwise, the register only where neither says, as for a load, whose
operands are integers whatever it loads. The readers are taken before
they move to the variable, since afterwards the value has none. Both
typings now share the one rule.

That exposed two gaps in what operations say about themselves. imul
and udiv built their nodes on a path of their own that never set
ConsumesInteger, so an integer multiply said nothing; and a structured
or raw load's address was not counted as an integer read, so the
`ishl` that made it said nothing either. compute_reduce's
`int t2 = (float)(max(count, 1))` was the same defect the other way
about, and is a float now.

+++ HlslDecompiler.Tests/ShaderSources/cs_4_1/compute_reduce.fx
@@ -11,6 +11,6 @@ void main(uint3 sv_dispatchthreadid : SV_DispatchThreadID)
 	for (int t1 = 0; t1 < count; t1 = t1 + 1) {
 		t0 = input[t1] * k + t0;
 	}
-	int t2 = (float)(max(count, 1));
+	float t2 = (float)(max(count, 1));
 	output[sv_dispatchthreadid.x] = t0 / t2;
 }
